modifier carte bancaire

+++ WattElse-Entity/src/main/java/fr/eql/ai116/proj2/tim/entity/BankCard.java
@@ -44,9 +44,11 @@ public class BankCard implements Serializable {
         return expirationDate;
     }
 
-    public long getCvvNumber() {
+    public int getCvvNumber() {
         return cvvNumber;
     }
+
+
 
 
     //Setters/////
